scraper/requirements.txt
selenium
kafka-python
pandas
